Unify the ast/eager tests. Fixed bug with lambda being popped off the value stack to early
