+++ python/cudaq/kernel/ast_bridge.py
@@ -1123,8 +1123,9 @@ class PyASTBridge(ast.NodeVisitor):
             if node.func.value.id in ['h', 'x', 'y', 'z', 's', 't'
                                      ] and node.func.attr == 'ctrl':
                 target = self.popValue()
+                # Should be len(node.args) - 1 controls (-1 for the target)
                 controls = [
-                    self.popValue() for i in range(len(self.valueStack))
+                    self.popValue() for i in range(len(node.args)-1)
                 ]
                 negatedControlQubits = None
                 if len(self.controlNegations):
