+++ python/tests/builder/test_SpinOperator.py
