+++ python/tests/kernel/test_library_kernels_eager.py
@@ -12,7 +12,16 @@ import numpy as np
 
 import cudaq
 
-def test_fswap_lib_kernel():
+@pytest.fixture(autouse=True)
+def do_something():
+    if os.getenv("CUDAQ_PYTEST_EAGER_MODE") == 'OFF':
+        cudaq.enable_jit()
+    yield
+    
+    if cudaq.jit_enabled(): cudaq.__clearKernelRegistries()
+    cudaq.disable_jit()
+
+def test_internal_library_kernels():
     from cudaq.lib import fermionic_swap
 
     angle = 0.2
@@ -49,20 +58,20 @@ def test_fswap_lib_kernel():
                       atol=1e-3)
 
 
-def test_givens_lib_kernel():
     from cudaq.lib import givens
 
     angle = 0.2
     c = np.cos(angle)
     si = np.sin(angle)
 
+    # PYTEST Error, baz fails
     @cudaq.kernel
-    def baz(angle: float):
+    def baz2(angle: float):
         q = cudaq.qlist(2)
         x(q[0])
         givens(angle, q[0], q[1])
 
-    ss_01 = cudaq.get_state(baz, angle)
+    ss_01 = cudaq.get_state(baz2, angle)
     print(ss_01)
     assert np.isclose(ss_01[1], c, 1e-3)
     assert np.isclose(ss_01[2], -si, 1e-3)
