+++ python/tests/kernel/test_sampling_eager.py
@@ -10,20 +10,29 @@ import os
 
 import pytest
 import numpy as np
+from typing import Callable 
 
 import cudaq
 
+@pytest.fixture(autouse=True)
+def do_something():
+    if os.getenv("CUDAQ_PYTEST_EAGER_MODE") == 'OFF':
+        cudaq.enable_jit()
+    yield
+    if cudaq.jit_enabled(): cudaq.__clearKernelRegistries()
+    cudaq.disable_jit()
 
 def test_simple_sampling_ghz():
     """Test that we can build a very simple kernel and sample it."""
 
     @cudaq.kernel
-    def simple(numQubits):
+    def simple(numQubits:int):
         qubits = cudaq.qvector(numQubits)
         h(qubits.front())
         for i, qubit in enumerate(qubits.front(numQubits - 1)):
             x.ctrl(qubit, qubits[i + 1])
 
+    print(simple)
     counts = cudaq.sample(simple, 10)
     assert len(counts) == 2
     assert '0' * 10 in counts and '1' * 10 in counts
@@ -37,7 +46,7 @@ def test_simple_sampling_qpe():
     """Test that we can build up a set of kernels, compose them, and sample."""
 
     @cudaq.kernel
-    def iqft(qubits):
+    def iqft(qubits:cudaq.qview):
         N = qubits.size()
         for i in range(N // 2):
             swap(qubits[i], qubits[N - i - 1])
@@ -51,15 +60,15 @@ def test_simple_sampling_qpe():
         h(qubits[N - 1])
 
     @cudaq.kernel
-    def tGate(qubit):
+    def tGate(qubit:cudaq.qubit):
         t(qubit)
 
     @cudaq.kernel
-    def xGate(qubit):
+    def xGate(qubit:cudaq.qubit):
         x(qubit)
 
     @cudaq.kernel
-    def qpe(nC, nQ, statePrep, oracle):
+    def qpe(nC:int, nQ:int, statePrep:Callable[[cudaq.qubit], None], oracle:Callable[[cudaq.qubit], None]):
         q = cudaq.qvector(nC + nQ)
         countingQubits = q.front(nC)
         stateRegister = q.back()
